refactor: remove unused component, tidy up code
