--- src/components/Attribution/index.jsx
@@ -1,15 +1,0 @@
-import { SWAPR_LINK } from '../../constants';
-
-const Attribution = () => (
-  <p className="attribution">
-    <a href="https://github.com/SwaprDAO/swapr-info" rel="noopener noreferrer" target="_blank">
-      Github
-    </a>{' '}
-    |{' '}
-    <a href={SWAPR_LINK} rel="noopener noreferrer" target="_blank">
-      Swapr
-    </a>{' '}
-  </p>
-);
-
-export default Attribution;
