+++ src/components/DropdownBasicSelect/index.jsx
@@ -1,18 +1,19 @@
+import PropTypes from 'prop-types';
 import { useRef, useState } from 'react';
 import { useClickAway } from 'react-use';
 
 import { StyledIcon } from '..';
 import { Typography } from '../../Theme';
 import { useDarkModeManager } from '../../contexts/LocalStorage';
-import { AutoColumn } from '../Column';
 import Icon from '../Icon';
 import Row, { RowBetween } from '../Row';
 import { ArrowStyled, Dropdown, Wrapper } from './styled';
 
-export default function DropdownBasicSelect({ options, active, disabled, setActive, color, width = null }) {
+const DropdownBasicSelect = ({ options, active, disabled, setActive, color, width = null }) => {
+  const [showDropdown, toggleDropdown] = useState(false);
+
   const darkMode = useDarkModeManager();
 
-  const [showDropdown, toggleDropdown] = useState(false);
   const dropdownRef = useRef(null);
   const containerRef = useRef(null);
 
@@ -44,32 +45,41 @@ export default function DropdownBasicSelect({ options, active, disabled, setActi
       {showDropdown && (
         <Dropdown>
           <div ref={dropdownRef}>
-            <AutoColumn gap={'20px'}>
-              {Object.keys(options).map((key, index) => {
-                let option = options[key];
-                return (
-                  option !== active && (
-                    <Row
-                      onClick={() => {
-                        toggleDropdown(!showDropdown);
-                        setActive(option);
-                      }}
-                      key={index}
+            {Object.keys(options).map((key, index) => {
+              let option = options[key];
+              return (
+                option !== active && (
+                  <Row
+                    onClick={() => {
+                      toggleDropdown(!showDropdown);
+                      setActive(option);
+                    }}
+                    key={index}
+                  >
+                    <Typography.SmallHeader
+                      color={'text8'}
+                      sx={{ display: 'flex', ':hover': { color: darkMode ? '#EBE9F8' : '#464366' } }}
                     >
-                      <Typography.SmallHeader
-                        color={'text8'}
-                        sx={{ display: 'flex', ':hover': { color: darkMode ? '#EBE9F8' : '#464366' } }}
-                      >
-                        {option}
-                      </Typography.SmallHeader>
-                    </Row>
-                  )
-                );
-              })}
-            </AutoColumn>
+                      {option}
+                    </Typography.SmallHeader>
+                  </Row>
+                )
+              );
+            })}
           </div>
         </Dropdown>
       )}
     </Wrapper>
   );
-}
+};
+
+DropdownBasicSelect.propTypes = {
+  options: PropTypes.any,
+  active: PropTypes.string,
+  disabled: PropTypes.bool,
+  setActive: PropTypes.func,
+  color: PropTypes.string,
+  width: PropTypes.string,
+};
+
+export default DropdownBasicSelect;
