+++ src/components/FarmingList/ListItem.jsx
@@ -1,7 +1,6 @@
-import React from 'react';
+import PropTypes from 'prop-types';
 import { ExternalLink } from 'react-feather';
 import { Flex } from 'rebass';
-import styled from 'styled-components';
 
 import { USD } from '@swapr/sdk';
 
@@ -15,47 +14,7 @@ import DoubleTokenLogo from '../DoubleLogo';
 import FormattedName from '../FormattedName';
 import Link, { InternalListLink } from '../Link';
 import TokenLogo from '../TokenLogo';
-
-export const DashGrid = styled.div`
-  display: grid;
-  grid-gap: 1em;
-  grid-template-columns: 0.5fr 0.9fr 0.5fr 0.4fr 0.2fr;
-  grid-template-areas: 'pair tvl apy rewardTokens link';
-  padding: 10px 36px;
-  height: 100px;
-
-  > * {
-    justify-content: flex-end;
-
-    :first-child {
-      justify-content: flex-start;
-      text-align: left;
-    }
-  }
-
-  @media screen and (min-width: 680px) {
-    display: grid;
-    grid-gap: 1em;
-    grid-template-columns: 0.6fr 0.6fr 0.6fr 0.6fr 0.4fr 0.2fr;
-    grid-template-areas: 'pair tvl yield apy rewardTokens link';
-
-    > * {
-      justify-content: flex-end;
-      width: 100%;
-
-      &:first-child {
-        justify-content: flex-start;
-      }
-    }
-  }
-
-  @media screen and (min-width: 1081px) {
-    display: grid;
-    grid-gap: 0.5em;
-    grid-template-columns: 0.2fr 1fr 1fr 1fr 1fr 1fr 0.5fr 0.5fr;
-    grid-template-areas: 'index pair tvl yield apy rewardTokens owner link';
-  }
-`;
+import { DashGrid } from './styled';
 
 const FlexText = ({ area, justifyContent, flexDirection, color, children }) => (
   <Flex area={area} justifyContent={justifyContent} flexDirection={flexDirection || 'row'}>
@@ -65,7 +24,7 @@ const FlexText = ({ area, justifyContent, flexDirection, color, children }) => (
   </Flex>
 );
 
-export default function ListItem({ campaign, index, nativeCurrencyPrice }) {
+const ListItem = ({ campaign, index, nativeCurrencyPrice }) => {
   const below600 = useIsBelowPx(600);
   const below680 = useIsBelowPx(680);
   const below740 = useIsBelowPx(740);
@@ -194,4 +153,12 @@ export default function ListItem({ campaign, index, nativeCurrencyPrice }) {
   } else {
     return null;
   }
-}
+};
+
+ListItem.propTypes = {
+  campaign: PropTypes.object,
+  index: PropTypes.number,
+  nativeCurrencyPrice: PropTypes.number,
+};
+
+export default ListItem;
