+++ src/components/FormattedName/index.jsx
@@ -1,24 +1,8 @@
+import PropTypes from 'prop-types';
 import { useState } from 'react';
-import styled from 'styled-components';
 
 import { Tooltip } from '../QuestionHelper';
-
-const TextWrapper = styled.div`
-  position: relative;
-  margin-left: ${({ margin }) => margin && '4px'};
-  color: ${({ theme, link }) => (link ? theme.swaprLink : theme.text1)};
-  font-size: ${({ fontSize }) => fontSize};
-  flex-basis: ${({ flexBasis }) => (flexBasis ? flexBasis : '')};
-  text-align: ${({ textAlign }) => (textAlign ? textAlign : '')};
-
-  :hover {
-    cursor: ${({ link }) => (link ? 'pointer' : 'default')};
-  }
-
-  @media screen and (max-width: 600px) {
-    font-size: ${({ adjustSize }) => adjustSize && '12px'};
-  }
-`;
+import { TextWrapper } from './styled';
 
 const FormattedName = ({
   text,
@@ -58,6 +42,14 @@ const FormattedName = ({
       {text}
     </TextWrapper>
   );
+};
+
+FormattedName.propTypes = {
+  text: PropTypes.string,
+  maxCharacters: PropTypes.number,
+  margin: PropTypes.bool,
+  adjustSize: PropTypes.bool,
+  fontSize: PropTypes.string,
 };
 
 export default FormattedName;
