+++ src/components/PairList/index.jsx
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
+import PropTypes from 'prop-types';
 import { useState, useEffect } from 'react';
-import { Box, Flex, Text } from 'rebass';
-import styled from 'styled-components';
+import { Flex } from 'rebass';
 
 import { Typography } from '../../Theme';
 import { Divider } from '../../components';
@@ -16,71 +16,9 @@ import LocalLoader from '../LocalLoader';
 import PageButtons from '../PageButtons';
 import Panel from '../Panel';
 import QuestionHelper from '../QuestionHelper';
+import { ClickableText, DashGrid, FeeText, List } from './styled';
 
 dayjs.extend(utc);
-
-const List = styled(Box)`
-  -webkit-overflow-scrolling: touch;
-`;
-
-const DashGrid = styled.div`
-  display: grid;
-  grid-gap: 1em;
-  grid-template-columns: 100px 1fr;
-  grid-template-areas: 'name vol';
-  padding: 0 20px;
-
-  > * {
-    justify-content: flex-end;
-
-    :first-child {
-      justify-content: flex-start;
-      text-align: left;
-    }
-  }
-
-  @media screen and (min-width: 680px) {
-    padding: 0 36px;
-    display: grid;
-    grid-gap: 1em;
-    grid-template-columns: 180px 1fr 1fr 1fr;
-    grid-template-areas: 'name symbol liq vol';
-
-    > * {
-      justify-content: flex-end;
-      width: 100%;
-
-      &:first-child {
-        justify-content: flex-start;
-      }
-    }
-  }
-
-  @media screen and (min-width: 1080px) {
-    padding: 0 36px;
-    display: grid;
-    grid-gap: 0.5em;
-    grid-template-columns: 0.1fr 1.5fr 0.6fr 1fr 1fr 1fr 1fr;
-    grid-template-areas: 'index name symbol liq vol price change';
-  }
-`;
-
-const ClickableText = styled(Text)`
-  color: ${({ theme }) => theme.text1};
-  &:hover {
-    cursor: pointer;
-    opacity: 0.6;
-  }
-  text-align: end;
-  user-select: none;
-`;
-
-const FeeText = styled.div`
-  background-color: ${({ theme }) => theme.bg8};
-  border-radius: 6px;
-  margin-left: 6px;
-  padding: 4px 6px;
-`;
 
 const FlexText = ({ area, justifyContent, color, children }) => (
   <Flex area={area} justifyContent={justifyContent}>
@@ -105,7 +43,7 @@ const FIELD_TO_VALUE = {
   [SORT_FIELD.FEES]: 'oneDayVolumeUSD',
 };
 
-export default function PairList({ pairs, maxItems = 10 }) {
+const PairList = ({ pairs, maxItems = 10 }) => {
   const below600 = useIsBelowPx(600);
   const below680 = useIsBelowPx(680);
   const below1080 = useIsBelowPx(1080);
@@ -327,4 +265,11 @@ export default function PairList({ pairs, maxItems = 10 }) {
       />
     </>
   );
-}
+};
+
+PairList.propTypes = {
+  pairs: PropTypes.any,
+  maxItems: PropTypes.number,
+};
+
+export default PairList;
