+++ src/components/Popover/index.tsx
@@ -1,80 +1,10 @@
 import type { Placement } from '@popperjs/core';
 import Portal from '@reach/portal';
-import { transparentize } from 'polished';
 import React, { useState } from 'react';
 import { usePopper } from 'react-popper';
-import styled from 'styled-components';
 
 import useInterval from '../../hooks';
-
-const PopoverContainer = styled.div<{ show: boolean }>`
-  z-index: 9999;
-
-  visibility: ${(props) => (props.show ? 'visible' : 'hidden')};
-  opacity: ${(props) => (props.show ? 1 : 0)};
-  transition: visibility 150ms linear, opacity 150ms linear;
-
-  background: ${({ theme }) => theme.bg2};
-  border: 1px solid ${({ theme }) => theme.bg3};
-  box-shadow: 0 4px 8px 0 ${({ theme }) => transparentize(0.9, theme.shadow1)};
-  color: ${({ theme }) => theme.text2};
-  border-radius: 8px;
-`;
-
-const ReferenceElement = styled.div`
-  display: inline-block;
-`;
-
-const Arrow = styled.div`
-  width: 8px;
-  height: 8px;
-  z-index: 9998;
-
-  ::before {
-    position: absolute;
-    width: 8px;
-    height: 8px;
-    z-index: 9998;
-
-    content: '';
-    border: 1px solid ${({ theme }) => theme.bg3};
-    transform: rotate(45deg);
-    background: ${({ theme }) => theme.bg2};
-  }
-
-  &.arrow-top {
-    bottom: -5px;
-    ::before {
-      border-top: none;
-      border-left: none;
-    }
-  }
-
-  &.arrow-bottom {
-    top: -5px;
-    ::before {
-      border-bottom: none;
-      border-right: none;
-    }
-  }
-
-  &.arrow-left {
-    right: -5px;
-
-    ::before {
-      border-bottom: none;
-      border-left: none;
-    }
-  }
-
-  &.arrow-right {
-    left: -5px;
-    ::before {
-      border-right: none;
-      border-top: none;
-    }
-  }
-`;
+import { Arrow, PopoverContainer, ReferenceElement } from './styled';
 
 export interface PopoverProps {
   content: React.ReactNode;
@@ -83,7 +13,7 @@ export interface PopoverProps {
   placement?: Placement;
 }
 
-export default function Popover({ content, show, children, placement = 'auto' }: PopoverProps) {
+const Popover = ({ content, show, children, placement = 'auto' }: PopoverProps) => {
   const [referenceElement, setReferenceElement] = useState<HTMLDivElement>(null);
   const [popperElement, setPopperElement] = useState<HTMLDivElement>(null);
   const [arrowElement, setArrowElement] = useState<HTMLDivElement>(null);
@@ -114,4 +44,6 @@ export default function Popover({ content, show, children, placement = 'auto' }:
       </Portal>
     </>
   );
-}
+};
+
+export default Popover;
