+++ src/components/QuestionHelper/index.tsx
@@ -1,34 +1,8 @@
 import React, { useCallback, useState } from 'react';
 import { HelpCircle as Question } from 'react-feather';
-import styled from 'styled-components';
 
 import Popover, { PopoverProps } from '../Popover';
-
-const QuestionWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0.2rem;
-  border: none;
-  background: none;
-  outline: none;
-  cursor: default;
-  border-radius: 36px;
-  background-color: ${({ theme }) => theme.bg2};
-  color: ${({ theme }) => theme.text2};
-
-  :hover,
-  :focus {
-    opacity: 0.7;
-  }
-`;
-
-const TooltipContainer = styled.div`
-  width: 228px;
-  padding: 0.6rem 1rem;
-  line-height: 150%;
-  font-weight: 400;
-`;
+import { QuestionWrapper, TooltipContainer } from './styled';
 
 interface TooltipProps extends Omit<PopoverProps, 'content'> {
   text: string;
@@ -38,7 +12,7 @@ export function Tooltip({ text, ...rest }: TooltipProps) {
   return <Popover content={<TooltipContainer>{text}</TooltipContainer>} {...rest} />;
 }
 
-export default function QuestionHelper({ text, disabled }: { text: string; disabled?: boolean }) {
+const QuestionHelper = ({ text, disabled }: { text: string; disabled?: boolean }) => {
   const [show, setShow] = useState<boolean>(false);
 
   const open = useCallback(() => setShow(true), [setShow]);
@@ -53,4 +27,6 @@ export default function QuestionHelper({ text, disabled }: { text: string; disab
       </Tooltip>
     </span>
   );
-}
+};
+
+export default QuestionHelper;
