+++ src/components/Search/styled.jsx
@@ -1,0 +1,125 @@
+import { X, Search as SearchIcon } from 'react-feather';
+import styled from 'styled-components';
+
+import Row from '../Row';
+
+const Container = styled.div`
+  position: relative;
+
+  @media screen and (max-width: 600px) {
+    width: 100%;
+  }
+`;
+
+const Wrapper = styled.div`
+  display: flex;
+  position: relative;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-end;
+  padding: 12px 16px;
+  border-radius: 12px;
+  background: ${({ theme }) => theme.bg7};
+  border-bottom-right-radius: ${({ open }) => (open ? '0px' : '12px')};
+  border-bottom-left-radius: ${({ open }) => (open ? '0px' : '12px')};
+  width: 100%;
+  min-width: 300px;
+  box-sizing: border-box;
+  box-shadow: ${({ open }) =>
+    !open
+      ? '0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04) '
+      : 'none'};
+  @media screen and (max-width: 500px) {
+    background: ${({ theme }) => theme.bg7};
+    box-shadow: ${({ open }) =>
+      !open
+        ? '0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04) '
+        : 'none'};
+  }
+`;
+const Input = styled.input`
+  position: relative;
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+  background: none;
+  border: none;
+  outline: none;
+  width: 100%;
+  color: ${({ theme }) => theme.text10};
+  font-size: ${({ large }) => (large ? '20px' : '14px')};
+
+  ::placeholder {
+    color: ${({ theme }) => theme.text3};
+    font-size: 16px;
+  }
+
+  @media screen and (max-width: 640px) {
+    ::placeholder {
+      font-size: 14px;
+    }
+  }
+`;
+
+const SearchIconLarge = styled(SearchIcon)`
+  height: 20px;
+  width: 20px;
+  margin-right: 0.5rem;
+  position: absolute;
+  right: 10px;
+  :hover {
+    cursor: pointer;
+  }
+
+  color: ${({ theme }) => theme.text3};
+`;
+
+const CloseIcon = styled(X)`
+  height: 20px;
+  width: 20px;
+  margin-right: 0.5rem;
+  position: absolute;
+  right: 10px;
+  color: ${({ theme }) => theme.text3};
+  :hover {
+    cursor: pointer;
+  }
+`;
+
+const Menu = styled.div`
+  position: absolute;
+  z-index: 2;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  top: 40px;
+  max-height: 540px;
+  overflow: auto;
+  left: 0;
+  margin-bottom: 20px;
+  background: ${({ theme }) => theme.bg7};
+  border-bottom-right-radius: 12px;
+  border-bottom-left-radius: 12px;
+  box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04),
+    0px 24px 32px rgba(0, 0, 0, 0.04);
+  display: ${({ hide }) => hide && 'none'};
+`;
+
+const MenuItem = styled(Row)`
+  padding: 1rem;
+  font-size: 0.85rem;
+  & > * {
+    margin-right: 6px;
+  }
+  :hover {
+    cursor: pointer;
+    background-color: ${({ theme }) => theme.bg2};
+  }
+`;
+
+const Heading = styled(Row)`
+  padding: 1rem;
+  display: ${({ hide = false }) => hide && 'none'};
+`;
+
+export { Container, Wrapper, Input, SearchIconLarge, CloseIcon, Menu, MenuItem, Heading };
