+++ src/components/Select/styles.jsx
@@ -226,4 +226,51 @@ export const customStylesTime = {
   }),
 };
 
+const MenuLabel = styled.div`
+  display: flex;
+  align-items: center;
+  width: 100%;
+  justify-content: flex-start;
+  flex-direction: row;
+`;
+
+const LabelBox = styled.div``;
+
+const LogoBox = styled.div`
+  width: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  margin-right: 8px;
+`;
+
+const CustomMenu = styled.div`
+  background-color: white;
+  position: absolute;
+  border-radius: 16px;
+  box-shadow: 0 4px 8px 0 rgba(47, 128, 237, 0.1), 0 0 0 0.5px var(--c-zircon);
+  overflow: hidden;
+  padding: 0;
+  width: 180px;
+  z-index: 5;
+  margin-top: 10px;
+  padding-top: 36px;
+`;
+
+const FixedToggle = styled.div`
+  position: absolute;
+  height: 24px;
+  z-index: 10;
+  background-color: white;
+  width: 100%;
+  top: 8px;
+  display: flex;
+  align-items: center;
+  padding-left: 12px;
+  & > input {
+    margin-right: 8px;
+  }
+`;
+
 export default customStyles;
+export { MenuLabel, LogoBox, LabelBox, CustomMenu, FixedToggle };
