+++ src/components/Title/styled.jsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const TitleWrapper = styled.div`
+  text-decoration: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  z-index: 10;
+`;
+
+export { TitleWrapper };
