+++ src/components/TxnList/index.jsx
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
+import PropTypes from 'prop-types';
 import { useState, useEffect } from 'react';
-import { Box, Flex, Text } from 'rebass';
-import styled from 'styled-components';
+import { Flex } from 'rebass';
 
 import { Divider, EmptyCard } from '..';
 import { Typography } from '../../Theme';
@@ -15,98 +15,9 @@ import { ExternalListLink } from '../Link';
 import LocalLoader from '../LocalLoader';
 import PageButtons from '../PageButtons';
 import Panel from '../Panel';
+import { ClickableText, DashGrid, List, SortText } from './styled';
 
 dayjs.extend(utc);
-
-const List = styled(Box)`
-  -webkit-overflow-scrolling: touch;
-`;
-
-const DashGrid = styled.div`
-  display: grid;
-  grid-gap: 1em;
-  grid-template-columns: 20px 1fr 1fr;
-  grid-template-areas: 'txn value time';
-  padding: 0 20px;
-
-  > * {
-    justify-content: flex-end;
-    width: 100%;
-
-    &:first-child {
-      justify-content: flex-start;
-      text-align: left;
-      width: 100px;
-    }
-  }
-
-  @media screen and (min-width: 680px) {
-    padding: 0 36px;
-
-    > * {
-      &:first-child {
-        width: 180px;
-      }
-    }
-  }
-
-  @media screen and (min-width: 780px) {
-    padding: 0 36px;
-    grid-template-columns: 1.2fr 1fr 1fr 1fr 1fr;
-    grid-template-areas: 'txn value amountToken amountOther time';
-
-    > * {
-      &:first-child {
-        width: 180px;
-      }
-    }
-  }
-
-  @media screen and (min-width: 1080px) {
-    padding: 0 36px;
-    grid-template-columns: 1.2fr 1fr 1fr 1fr 1fr 1fr;
-    grid-template-areas: 'txn value amountToken amountOther account time';
-  }
-`;
-
-const ClickableText = styled(Text)`
-  color: ${({ theme }) => theme.text1};
-  user-select: none;
-  text-align: end;
-
-  &:hover {
-    cursor: pointer;
-    opacity: 0.6;
-  }
-
-  @media screen and (max-width: 640px) {
-    font-size: 14px;
-  }
-`;
-
-const SortText = styled.button`
-  padding: 4px 5px;
-  border: 1px solid;
-  border-color: ${({ theme }) => theme.bd1};
-  border-radius: 6px;
-  background-color: ${({ isActive, theme }) => (isActive ? theme.bg2 : 'transparent')};
-  color: ${({ isActive, theme }) => (isActive ? theme.text12 : theme.text7)};
-  text-transform: uppercase;
-
-  :hover {
-    cursor: pointer;
-  }
-
-  :hover > * {
-    color: ${({ theme }) => theme.text12};
-  }
-
-  & > div {
-    color: ${({ isActive, theme }) => (isActive ? theme.text12 : theme.text7)};
-  }
-
-  transition: background 200ms;
-`;
 
 const FlexText = ({ area, color, children }) => (
   <Typography.LargeText color={color || 'text1'} sx={{ gridArea: area, display: 'flex', alignItems: 'center' }}>
@@ -130,7 +41,7 @@ const TXN_TYPE = {
 
 const ITEMS_PER_PAGE = 10;
 
-function getTransactionType(event, symbol0, symbol1, short) {
+const getTransactionType = (event, symbol0, symbol1, short) => {
   const formattedS0 = symbol0?.length > 8 ? symbol0.slice(0, 7) + '...' : symbol0;
   const formattedS1 = symbol1?.length > 8 ? symbol1.slice(0, 7) + '...' : symbol1;
   switch (event) {
@@ -143,10 +54,9 @@ function getTransactionType(event, symbol0, symbol1, short) {
     default:
       return '';
   }
-}
+};
 
-// @TODO rework into virtualized list
-function TxnList({ transactions, symbol0Override, symbol1Override }) {
+const TransactionsList = ({ transactions, symbol0Override, symbol1Override }) => {
   const below1080 = useIsBelowPx(1080);
   const below780 = useIsBelowPx(780);
   const below680 = useIsBelowPx(680);
@@ -456,6 +366,12 @@ function TxnList({ transactions, symbol0Override, symbol1Override }) {
       />
     </>
   );
-}
+};
 
-export default TxnList;
+TransactionsList.propTypes = {
+  transactions: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
+  symbol0Override: PropTypes.string,
+  symbol1Override: PropTypes.string,
+};
+
+export default TransactionsList;
