+++ src/components/Warning/styled.jsx
@@ -1,0 +1,26 @@
+import { AlertTriangle } from 'react-feather';
+import styled from 'styled-components';
+
+const WarningWrapper = styled.div`
+  border-radius: 20px;
+  border: 1px solid #f82d3a;
+  background: rgba(248, 45, 58, 0.05);
+  padding: 1rem;
+  color: #f82d3a;
+  display: ${({ show }) => !show && 'none'};
+  margin: 0 2rem 2rem 2rem;
+  position: relative;
+
+  @media screen and (max-width: 800px) {
+    width: 80% !important;
+    margin-left: 5%;
+  }
+`;
+
+const StyledWarningIcon = styled(AlertTriangle)`
+  min-height: 20px;
+  min-width: 20px;
+  stroke: red;
+`;
+
+export { WarningWrapper, StyledWarningIcon };
